requirements.txt
absl-py 
accelerate==0.23.0
aiohttp 
aiosignal 
altair==5.0.1
appdirs 
arrow 
asttokens 
astunparse==1.6.3
async-timeout 
attrs 
backcall 
beautifulsoup4==4.12.2
binaryornot 
blinker==1.4
blis 
Bottleneck 
brotlipy 
cachetools 
catalogue 
certifi 
cffi 
chardet 
charset-normalizer 
click 
cmake==3.27.5
colorama 
comm 
# conda-content-trust 
# conda-package-handling 
# conda_package_streaming 
contourpy 
cookiecutter 
cryptography 
cupy 
cvxpy==1.3.2
cycler 
cymem 
Cython==0.29.35
dataclasses 
datasets 
debugpy 
decorator 
dill 
docker-pycreds==0.4.0
ecos==2.0.12
# en-core-web-lg 
# en-core-web-sm 
evaluate 
executing 
fastrlock==0.8
filelock 
flatbuffers==23.3.3
flit_core 
fonttools==4.25.0
frozenlist 
fsspec 
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.99.0
google-auth==2.17.3
google-auth-httplib2==0.1.1
google-auth-oauthlib>=0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.60.0
grpcio 
h5py==3.8.0
httplib2==0.22.0
huggingface-hub 
idna 
importlib-metadata 
ipykernel 
ipython 
jax==0.4.8
jedi 
Jinja2 
jinja2-time 
joblib 
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_client 
jupyter_core 
keras==2.12.0
Keras-Preprocessing 
keybert==0.7.0
kiwisolver 
langcodes 
Levenshtein==0.20.9
libclang==16.0.0
lit==16.0.6
lxml==4.9.2
mapping==0.1.6
Markdown 
markdown-it-py==2.2.0
MarkupSafe 
matplotlib 
matplotlib-inline 
mdurl==0.1.2
mkl-fft>=1.3.1
mkl-random 
mkl-service==2.4.0
ml-dtypes==0.1.0
mlxtend==0.22.0
mpmath==1.3.0
mstranslator==0.3.2
multidict 
multiprocess 
munkres==1.1.4
murmurhash 
nest-asyncio 
networkx==3.1
nltk 
numexpr 
numpy 
oauthlib 
openai==0.27.6
opt-einsum==3.3.0
osqp==0.6.3
packaging 
pandarallel==1.6.5
pandas==1.5.3
parso 
pathtools==0.1.2
pathy 
pexpect 
pickleshare 
Pillow==9.3.0
platformdirs 
plotly==5.16.1
pluggy 
ply==3.11
pooch 
portalocker==2.7.0
poyo 
preshed 
prompt-toolkit 
protobuf==3.20.3
psutil 
ptyprocess 
pure-eval 
pyarrow==11.0.0
pyasn1 
pyasn1-modules==0.2.8
pycosat 
pycparser 
pydantic 
pydeck==0.8.0
Pygments 
PyJWT 
Pympler==1.0.1
pyOpenSSL 
pyparsing 
PyQt5-sip==12.11.0
PySocks 
python-dateutil 
python-slugify 
pytorch-triton-rocm==2.0.2
pytz 
pytz-deprecation-shim==0.1.0.post0
PyYAML 
pyzmq 
qdldl==0.1.7.post0
rake-nltk==1.0.6
rapidfuzz==2.13.7
referencing==0.30.2
regex 
requests 
requests-oauthlib==1.3.0
responses 
rich==13.3.2
rpds-py==0.9.2
rsa 
ruamel.yaml 
ruamel.yaml.clib 
sacrebleu==2.3.1
sacremoses==0.0.43
safetensors==0.3.2
scikit-learn 
scipy==1.10.0
scs==3.2.3
seaborn==0.12.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.29.2
seqeval==1.2.2
setproctitle==1.3.2
setuptools-rust==1.7.0
shellingham 
simpletransformers==0.64.3
sip 
six 
smart-open 
smmap==5.0.0
sortedcontainers 
soupsieve==2.5
spacy 
spacy-legacy 
spacy-loggers 
srsly 
stack-data 
streamlit==1.25.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorboard-plugin-wit 
tensorboardX==2.6.2.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
text-unidecode 
thinc 
threadpoolctl 
tokenizers 
toml 
tomli==2.0.1
toolz
tornado 
tqdm 
traitlets 
transformers==4.33.3
triton==2.0.0
typer 
typing_extensions 
tzdata==2023.3
tzlocal==4.3.1
Unidecode 
uritemplate==4.1.1
urllib3 
validators==0.21.2
varname==0.11.0
wandb==0.15.8
wasabi 
watchdog==3.0.0
wcwidth 
Werkzeug 
wrapt==1.14.1
xxhash 
yarl 
zipp 
zstandard 
